requirements.txt
google-cloud-bigquery==0.29.0
google-cloud-storage==1.7.0
